+++ mountains/views/__init__.py
@@ -1,0 +1,3 @@
+from .views import *
+from .detail_views import *
+from .review_views import *
